% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{check_balance}
\alias{check_balance}
\title{Vérifie l'équilibre des 0 et 1 dans chaque ligne et colonne}
\usage{
check_balance(grid)
}
\arguments{
\item{grid}{Une matrice Takuzu}
}
\value{
TRUE si chaque ligne/colonne contient au plus la moitié de 0 ou de 1
}
\description{
Vérifie l'équilibre des 0 et 1 dans chaque ligne et colonne
}
