% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{check_no_na}
\alias{check_no_na}
\title{Vérifie que la grille ne contient pas de valeurs manquantes (NA)}
\usage{
check_no_na(grid)
}
\arguments{
\item{grid}{Une matrice Takuzu}
}
\value{
TRUE si aucune case n'est vide, FALSE sinon
}
\description{
Vérifie que la grille ne contient pas de valeurs manquantes (NA)
}
