% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{check_no_triplets}
\alias{check_no_triplets}
\title{Vérifie l'absence de triplets (000 ou 111) dans une grille Takuzu}
\usage{
check_no_triplets(grid)
}
\arguments{
\item{grid}{Une matrice Takuzu avec des 0, 1 ou NA}
}
\value{
TRUE si aucun triplet n'est présent, FALSE sinon
}
\description{
Vérifie l'absence de triplets (000 ou 111) dans une grille Takuzu
}
