% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{check_unique_rows_cols}
\alias{check_unique_rows_cols}
\title{Vérifie que toutes les lignes et colonnes sont uniques}
\usage{
check_unique_rows_cols(grid)
}
\arguments{
\item{grid}{Une matrice Takuzu}
}
\value{
TRUE si toutes les lignes et colonnes sont différentes (ignorant les lignes/colonnes incomplètes)
}
\description{
Vérifie que toutes les lignes et colonnes sont uniques
}
