% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid.R
\name{generer_takuzu_jouable}
\alias{generer_takuzu_jouable}
\title{Génère une grille Takuzu valide, puis la rend jouable en cachant des cases}
\usage{
generer_takuzu_jouable(taille = 8, proportion_visible = 0.3)
}
\arguments{
\item{taille}{Taille de la grille (4, 6 ou 8)}

\item{proportion_visible}{Proportion de cases visibles (ex: 0.3)}
}
\value{
Une liste avec deux éléments :
\describe{
\item{grille_visible}{Grille partiellement remplie (NA pour les cases cachées)}
\item{solution}{Grille complète}
}
}
\description{
Génère une grille Takuzu valide, puis la rend jouable en cachant des cases
}
